Force Windows to use Samba for file shares
Since Ansible is struggling with permissions issues on VirtualBox-mounted
file shares, we should force file sharing to use Samba. Samba (SMB) is
similar to NFS and is more performant that VBox shares, but it also has
better permissions management and allows for provisioning to flow through
to execution.

Fixes wpengine/hgv#96