Don't Overwrite Constant
If multiple HGV installations exist on the host, commands like `vagrant
global-status` attempt to parse all of the Vagrantfiles. This can cause
problems since all of them declare `VAGRANTFILE_API_VERSION`. By problems,
I mean this will trigger a Ruby warning that mucks with proper output.

To better protect against conflicts, we should check to see if the
constant is defined before we set it so we can avoid stumbling over
multiple installations co-located on a machine.