Update the S3 Uploads plugin
The ZIP file location provided by GitHub isn't to the actual file,
but redirects to a CDN-hosted file stream. This causes provisioning
to throw an error about the file not existing, resulting in the plugin
not getting installed on sites as intended. This fixes that by using the
CDN URL to the filestream.

This also bumps the version to the latest release of 1.1.0.